+++ src/bindings/bnd_embedded_file.h
@@ -23,10 +23,9 @@ public:
   BND_File3dmEmbeddedFile(ON_EmbeddedFile* ef, const ON_ModelComponentReference* compref);
 
   static BND_File3dmEmbeddedFile* Read(const std::wstring& f);
-  static BND_File3dmEmbeddedFile* Read2(const std::string& f);
-//#if defined(ON_WASM_COMPILE)
+#if defined(ON_WASM_COMPILE)
   static BND_File3dmEmbeddedFile* WasmFromByteArray(std::string buffer);
-//#endif
+#endif
 
   std::wstring GetFilename(void) const;
   void SetFilename(std::wstring filename){ _ef->SetFilename(filename.c_str()); };